refactor wallet transactions
